% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeastSquares.R
\name{coef.linreg}
\alias{coef.linreg}
\title{Implementation of coef function using linreg , an alternative for lm.coef.}
\usage{
\method{coef}{linreg}(x = mod_object, ...)
}
\arguments{
\item{x}{object of linreg}

\item{...}{optional parameter}
}
\value{
Named Vector of Coefficients
}
\description{
Implementation of coef function using linreg , an alternative for lm.coef.
}
\examples{
coef(linreg(formula = Petal.Length ~ Species, data = iris))

}
