% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeastSquares.R
\name{linreg}
\alias{linreg}
\title{Implementation of Liner Regression function linreg, an alternative for lm model.
Authors: Maria Treesa Sebastian(marse306), Brian Masinde(brima748), Omkar(omkbh878)}
\arguments{
\item{formula}{as formula}

\item{data}{as dataframe}
}
\value{
computedvalues  an object of type s3 class.
}
\description{
Implementation of Liner Regression function linreg, an alternative for lm model.
Authors: Maria Treesa Sebastian(marse306), Brian Masinde(brima748), Omkar(omkbh878)
}
\examples{
linreg(formula = Petal.Length ~ Species, data = iris)


}
