% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QR.R
\name{linreg_QR}
\alias{linreg_QR}
\title{Function to fit linear model by way of QR decomposition
Authors: Brian Masinde (brima748), Maria Treesa Sebastian (marse306), Omkar(omkbh878)}
\usage{
linreg_QR(formula, data)
}
\arguments{
\item{formula}{as formula}

\item{data}{as data frame}
}
\description{
Function to fit linear model by way of QR decomposition
Authors: Brian Masinde (brima748), Maria Treesa Sebastian (marse306), Omkar(omkbh878)
}
