% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeastSquares.R
\name{pred.linreg}
\alias{pred.linreg}
\title{Implementation of pred function using linreg , an alternative for lm.pred.}
\usage{
\method{pred}{linreg}(x, ...)
}
\arguments{
\item{x}{as object of linreg}
}
\value{
Vector of Fitted values
}
\description{
Implementation of pred function using linreg , an alternative for lm.pred.
}
\examples{
pred(linreg(formula = Petal.Length ~ Species, data = iris))

}
