% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeastSquares.R
\name{print.linreg}
\alias{print.linreg}
\title{Implementation of Print function using linreg , an alternative for lm.print.}
\usage{
\method{print}{linreg}(x, ...)
}
\arguments{
\item{x}{as object of linreg}
}
\value{
Nothing
}
\description{
Implementation of Print function using linreg , an alternative for lm.print.
}
\examples{
print(linreg(formula = Petal.Length ~ Species, data = iris))

}
