% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeastSquares.R
\name{resid.linreg}
\alias{resid.linreg}
\title{Implementation of resid function using linreg , an alternative for lm.resid.}
\usage{
\method{resid}{linreg}(x, ...)
}
\arguments{
\item{x}{as object of linreg}
}
\value{
Vector of residuals
}
\description{
Implementation of resid function using linreg , an alternative for lm.resid.
}
\examples{
resid(linreg(formula = Petal.Length ~ Species, data = iris))

}
