% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeastSquares.R
\name{resid}
\alias{resid}
\title{Generic resid function}
\usage{
resid(x, ...)
}
\arguments{
\item{x}{as s3 object}

\item{...}{optional parameter}
}
\value{
nothing
}
\description{
Generic resid function
}
