% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeastSquares.R
\name{summary.linreg}
\alias{summary.linreg}
\title{Implementation of summary function using linreg , an alternative for lm.summary.}
\usage{
\method{summary}{linreg}(x, ...)
}
\arguments{
\item{x}{as object of linreg}
}
\value{
summary of linreg with formula data, coefficient matrix and residual standard error
}
\description{
Implementation of summary function using linreg , an alternative for lm.summary.
}
\examples{
summary(linreg(formula = Petal.Length ~ Species, data = iris))

}
