% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeastSquares.R
\name{summary}
\alias{summary}
\title{Generic summary function}
\usage{
summary(x, ...)
}
\arguments{
\item{x}{as object}

\item{...}{optional parameter}
}
\value{
nothing
}
\description{
Generic summary function
}
