% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Linreq}
\value{
Linear Regresion
}
\description{
Linreq
}
\section{Fields}{

\describe{
\item{\code{x}}{matrix.}

\item{\code{y}}{numeric.}

\item{\code{hat_B}}{matrix.}

\item{\code{hat_y}}{matrix.}

\item{\code{hat_e}}{matrix.}

\item{\code{df}}{integer.}

\item{\code{hat_var_2}}{matrix.}

\item{\code{hat_var_hat_B}}{matrix.}

\item{\code{data}}{data.frame.}

\item{\code{formula}}{formula.}

\item{\code{t_B}}{numeric.}
}}


\examples{
data(iris)
mod_object <- lm(Petal.Length~Species, data = iris)
print(mod_object)
}
