% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg}
\alias{linreg}
\title{Initialize the linreg Object}
\arguments{
\item{formula}{A formula specifying the linear regression model (e.g., y ~ x).}

\item{data}{A data frame containing the data for fitting the model.}
}
\value{
An object with the calculated coefficients, residuals, fitted values,
and other related statistics.

A numeric vector of residuals (ˆe).

A numeric vector of fitted values.

A numeric vector of regression coefficients.
}
\description{
Initializes an object by performing QR decomposition
and calculating the regression coefficients, fitted values, residuals,
and residual variance.

Prints the formula used in the model and the regression coefficients.

Shows two plots: Residuals vs. Fitted values plot
and Scale-Location plot.

Returns the residuals (errors) of the regression model.

Returns the predicted values (fitted values) of the model.

Returns the regression coefficients (beta) of the model.

Summarizes the regression model by providing the coefficients,
standard errors, t-values, p-values, residual standard error, and degrees of freedom.
}
\details{
Linear Regression Using Reference Class

A function in form of class for fitting a linear regression model using
QR decomposition. Also supports basic regression methods like fitting the
model, printing the results, plotting diagnostics, and returning residuals,
predicted values, and coefficients.
}

