% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeastSquares.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Implementation of plot function using linreg , an alternative for lm.plot.}
\usage{
\method{plot}{linreg}(x, ...)
}
\arguments{
\item{x}{as object of linreg}

\item{...}{optional parameter}
}
\value{
Two plots  Residual vs Fitted  and  Scale-Location
}
\description{
Implementation of plot function using linreg , an alternative for lm.plot.
}
